Moved GOCachedObject to the model subdirectory
